% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_a.R
\name{extract_a}
\alias{extract_a}
\title{Agricultural Data}
\usage{
extract_a()
}
\value{
a df with all the agricultural data available in siap since 1980
}
\description{
Extract agricultural data available in siap since 1980
}
\examples{
agricultural_dataframe <- extract_a()
}
