% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_g.R
\name{extract_g}
\alias{extract_g}
\title{Cattle Data}
\usage{
extract_g()
}
\value{
a df with all the cattle data available in siap since 1980
}
\description{
Extract cattle data available in siap since 1980
}
\examples{
cattle_dataframe <- extract_g()
}
